REN-4398 Remove Pipfiles from gitignore
After a review on PR#4 REN-4398 is concluded that this repo is a
template for all looker projects. By keeping pip files in the gitignore
I am assuming that no python code will be introduced to the users of
this template. This is a bad assumption because they could very will
want to add something and it wont be tracked because of the gitignore.
There for these were removed.